project: flowchart: add method for file reading
method tries to read from file values.conf the first
five lines.
If the file isn't found, it just gives back
'FILENOTFOUND' as word to play Hangman with.

If the file is present, it reads the first five
lines, decides randomly which line the word is
choosen from.
Splits the choosen line into single words on each
space and again randomly choses the word from this
line to give back to the calling method